moved declaration of communications related defaults
